+++ src/storm-conv/settings/modules/ConversionOutputSettings.h
@@ -30,6 +30,21 @@ namespace storm {
                  */
                 std::string getJaniOutputFilename() const;
                 
+                /*!
+                 * Retrieves whether the output should be in the Prism format
+                 */
+                bool isPrismOutputSet() const;
+                
+                /*!
+                 * Retrieves whether an output filename for the prism file was specified
+                 */
+                bool isPrismOutputFilenameSet() const;
+
+                /*!
+                 * Retrieves the name of the prism output (if specified)
+                 */
+                std::string getPrismOutputFilename() const;
+                
                 bool check() const override;
                 void finalize() override;
 
@@ -41,6 +56,8 @@ namespace storm {
             private:
                 // Define the string names of the options as constants.
                 static const std::string janiOutputOptionName;
+                // Define the string names of the options as constants.
+                static const std::string prismOutputOptionName;
             };